Move some things back from applyFixes into the after hook
