+++ lib/applyFixes.js
@@ -3,8 +3,6 @@ const Traverser = require('eslint/lib/shared/traverser');
 const isSimpleObjectTree = require('./isSimpleObjectTree');
 const detectIndent = require('detect-indent');
 const SourceCodeFixer = require('eslint/lib/linter/source-code-fixer');
-const run = require('./run');
-const Promise = require('bluebird');
 const indentString = require('./indentString');
 const maybeApplyPrettier = require('./maybeApplyPrettier');
 
@@ -47,7 +45,7 @@ function getNodeIndent(node, sourceCode, byLastLine) {
   }
 }
 
-async function applyFixes(fixes) {
+function applyFixes(fixes) {
   const fixesByFileName = {};
   let numFixedExpects = 0;
   for (const {
@@ -181,88 +179,7 @@ async function applyFixes(fixes) {
     }
   }
 
-  const fileNames = Object.keys(fixedSourceTextByFileName);
-  let firstError;
-  if (fileNames.length > 0) {
-    const isUnderVersionControlByFileName = {};
-    const hasUncommittedChangesByFileName = {};
-    const results = await Promise.map(
-      fileNames,
-      fileName => run('git', ['ls-files', '--error-unmatch', fileName]),
-      { concurrency: 5 }
-    );
-
-    for (const [i, result] of results.entries()) {
-      if (result[1]) {
-        firstError = firstError || result[1];
-      } else {
-        isUnderVersionControlByFileName[fileNames[i]] = result[0] === 0;
-      }
-    }
-    const [, err, dirtyFileNames] = await run(
-      'git',
-      ['diff-index', 'HEAD', '--name-only'],
-      {
-        bufferLines: true
-      }
-    );
-
-    firstError = firstError || err;
-    for (const fileName of dirtyFileNames) {
-      hasUncommittedChangesByFileName[fileName] = true;
-    }
-
-    // Prevent the test runner from forcefully exiting while we're prompting, but save the call args for replaying later:
-    const originalProcessExit = process.exit;
-    const processExitArgs = [];
-    process.exit = function(...args) {
-      processExitArgs.push(args);
-    };
-
-    // Replay process.exit calls that happened while we were prompting and writing files:
-    // eslint-disable-next-line no-inner-declarations
-    function replayAndRestoreProcessExit() {
-      process.exit = originalProcessExit;
-      for (const args of processExitArgs) {
-        process.exit(...args);
-      }
-    }
-
-    // Give the test runner a chance to complete its output before we prompt:
-    setImmediate(async function() {
-      console.log(
-        `unexpected-snapshot detected ${numFixedExpects} failing snapshot assertion${
-          numFixedExpects === 1 ? '' : 's'
-        } in ${fileNames.length} source file${
-          fileNames.length === 1 ? '' : 's'
-        }`
-      );
-      if (firstError) {
-        console.log(`Aborting due to error: ${firstError.stack}`);
-        processExitArgs.unshift([165]); // Fail with a oddball error code that can be detected by tests
-        return replayAndRestoreProcessExit();
-      }
-
-      const Promise = require('bluebird');
-      const fs = Promise.promisifyAll(require('fs'));
-
-      if (fileNames.length > 0) {
-        const results = await Promise.map(
-          fileNames,
-          fileName =>
-            fs.writeFileAsync(fileName, fixedSourceTextByFileName[fileName]),
-          { concurrency: 5 }
-        );
-
-        console.log(
-          `unexpected-snapshot: Wrote ${results.length} file${
-            results.length === 1 ? '' : 's'
-          }`
-        );
-      }
-      replayAndRestoreProcessExit();
-    });
-  }
+  return { numFixedExpects, fixedSourceTextByFileName };
 }
 
 module.exports = applyFixes;
